refact: CloseIcon 색상 값으로 분기하도록 수정
