+++ src/app/_components/_modules/SpeechBubble.tsx
@@ -1,6 +1,6 @@
 import classNames from 'classnames/bind';
 import styles from './SpeechBubble.module.scss';
-import CloseIconBlack from '@/app/_svgs/CloseIconBlack';
+import CloseIcon from '@/app/_svgs/CloseIcon';
 import { ToggleContext } from '@/app/_hooks/useToggleContext';
 import { ToggleType } from '@/app/_hooks/useToggle';
 
@@ -32,7 +32,7 @@ const SpeechBubble = ({ context, message, direction = 'up', hasCloseBtn = false 
         {message}
         {hasCloseBtn && (
           <button className={cx('close')} onClick={handleClose}>
-            <CloseIconBlack />
+            <CloseIcon />
           </button>
         )}
       </div>
