+++ src/app/_svgs/CloseIcon.tsx
@@ -1,3 +1,13 @@
+interface Props {
+  color?: 'black' | 'white';
+}
+
+const CloseIcon = ({ color = 'black' }: Props) => {
+  const isBlack = color === 'black';
+
+  return isBlack ? CloseIconBlack() : CloseIconWhite();
+};
+
 const CloseIconBlack = () => {
   return (
     <i>
@@ -24,4 +34,23 @@ const CloseIconBlack = () => {
   );
 };
 
-export default CloseIconBlack;
+const CloseIconWhite = () => {
+  return (
+    <i>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="18"
+        height="18"
+        viewBox="0 0 18 18"
+        fill="none"
+      >
+        <path
+          d="M17.1693 2.4785L15.5243 0.833496L9.0026 7.35516L2.48094 0.833496L0.835938 2.4785L7.3576 9.00016L0.835938 15.5218L2.48094 17.1668L9.0026 10.6452L15.5243 17.1668L17.1693 15.5218L10.6476 9.00016L17.1693 2.4785Z"
+          fill="white"
+        />
+      </svg>
+    </i>
+  );
+};
+
+export default CloseIcon;
